requirements.txt
django>=1.8.4,<1.9
djangorestframework>=3.2.4,<3.3
